requirements.txt
streamlit
pandas
numpy
matplotlib
tensorflow-cpu
scikit-learn